chore: Try fix tests
