+++ tests/conftest.py
@@ -1,8 +1,44 @@
+import os
+import subprocess
+import sys
 from pathlib import Path
 
 import pytest
+import xarray as xr
+
+root_path = Path(__file__).parent.parent
+sys.path.insert(1, os.path.join(root_path, "scripts"))
+from omDownloadInputs import download_input_files, downloads, remote  # noqa: E402
 
 
 @pytest.fixture(scope="session")
 def root_dir():
     return Path(__file__).parent.parent
+
+
+# Fixture to download and later remove all input files
+@pytest.fixture(scope="session")
+def input_files(root_dir):
+    download_input_files(root_path=root_dir, downloads=downloads, remote=remote)
+
+    input_folder = os.path.join(root_dir, "inputs")
+
+    downloaded_files = os.listdir(input_folder)
+
+    yield downloaded_files
+
+    for file in [i for i in downloaded_files if i != "README.md"]:
+        filepath = os.path.join(input_folder, file)
+        os.remove(filepath)
+
+
+@pytest.fixture(scope="function")
+def input_domain_xr(root_dir, input_files):
+    subprocess.run(["python", os.path.join(root_dir, "scripts/omCreateDomainInfo.py")], check=True)  # noqa: S603, S607
+
+    # Generated by scripts/omCreateDomainInfo.py
+    filepath_in_domain = os.path.join(root_dir, "inputs/om-domain-info.nc")
+
+    yield xr.load_dataset(filepath_in_domain)
+
+    os.remove(filepath_in_domain)
