+++ tests/test_om_prior.py
@@ -16,7 +16,7 @@ from openmethane_prior.omUtils import getenv, secsPerYear
 
 root_path = Path(__file__).parent.parent
 sys.path.insert(1, os.path.join(root_path, "scripts"))
-from omDownloadInputs import download_input_files, downloads, remote  # noqa: E402
+from omDownloadInputs import downloads, remote  # noqa: E402
 
 
 @pytest.fixture(scope="session")
@@ -38,34 +38,6 @@ def geom_xr(root_dir):
     cmaqExamplePath = getenv("CMAQ_EXAMPLE")
     geomFilePath = os.path.join(root_dir, cmaqExamplePath, getenv("GEO_EM"))
     return xr.open_dataset(geomFilePath)
-
-
-# Fixture to download and later remove all input files
-@pytest.fixture(scope="session")
-def input_files(root_dir):
-    download_input_files(root_path=root_dir, downloads=downloads, remote=remote)
-
-    input_folder = os.path.join(root_dir, "inputs")
-
-    downloaded_files = os.listdir(input_folder)
-
-    yield downloaded_files
-
-    for file in [i for i in downloaded_files if i != "README.md"]:
-        filepath = os.path.join(input_folder, file)
-        os.remove(filepath)
-
-
-@pytest.fixture(scope="session")
-def input_domain_xr(root_dir, input_files):
-    subprocess.run(["python", os.path.join(root_dir, "scripts/omCreateDomainInfo.py")], check=True)  # noqa: S603, S607
-
-    # Generated by scripts/omCreateDomainInfo.py
-    filepath_in_domain = os.path.join(root_dir, "inputs/om-domain-info.nc")
-
-    yield xr.load_dataset(filepath_in_domain)
-
-    os.remove(filepath_in_domain)
 
 
 @pytest.fixture(scope="session")
